Set up testdata structure
When `main-test.js` is compiled, it loads the entire app, then loads some testdata in JSON format, which it then starts putting in the db object. I'm putting some testdata together of a flight in the French Alps, located in `testdata/stjean` folder. A http server must be run from the repo root in order for this testdata to be available from the app.