Map, timeline and blog working
Testdata can now be displayed. All basic functionality is working: the blog shows messages with a media thumbnail and launches a modal on click, the timeline shows the time of messages with an icon, and the map shows the track and messages with text and icon.

`main-test.js` reads some sample data and will go to the next message within a fixed interval of time. It waits a few seconds before moving on.