Basic Telegram bot and stubs for crossbar.io
Crossbar.io provides a websocket router with RPC and pub/sub functionality, simplifying the design as I don't need a custom API endpoint but can just do RPC calls to different microservices that get routed through Crossbar. It can be used with asyncio and is set up using a class with methods resembling lifecycle hooks.

The Telegram bot works if provided with a Telegram API token. It uses the very extensive *python-telegram-bot* library, which is not async but does run in a separate thread and provides an insane amount of functionality.