Message db
The asyncio-based message Db abstraction layer is now able to insert and retrieve messages. The control flow to transcode media is also in place, only needs to be hooked to an asyncio subprocess ffmpeg call.

Testing is done by using postgres nested transactions, which is great because it can be done on the main database as any changes are rolled back automatically when testing is complete or exceptions happen.

Execute the module using `python3 -m backend.messages.db [--create] [--test]` to create database tables or run tests. It's kind of ghetto without any abstractions like testing framework, but I like running stuff on the commandline and the control I have by not letting extra abstraction layers like unittest or pytest interfere. Also, there's not much in terms of async testing frameworks yet.