Restart on crossbar disconnect
When crossbar disconnects for some reason (e.g. the router process stops), the event loop is stopped, the process exits and is restarted by supervisor. The telegram bot uses threads and waits for the next long polling to complete, which takes an amount of time equal to the timeout specified. In production, with webhooks, this should not be an issue.

Might make it even more robust by attaching an exception handler on the event loop that listens to *autobahn.wamp.exception.TransportLost* and stops the loop when that happens, but that might be overkill as *onDisconnect* probably gets called always.