Analytics module and aiohttp logging middleware
A separate analytics service ingests logs from another WAMP service. The site service has some aiohttp middleware that sends pageviews with basic information to this service. It can be used both for performance logging (time to render page), but its main purpose is analytics tracking. It sends a cookie containing a UUID to the browser, to be able to identify page visit flows.

The idea is to make a dashboard that shows relevant metrics. The basic framework for that is in place behind `/analytics`. Should be protected with basic http auth or similar system.