Show message load error on frontpage
This change is in line with trying to build more robustness and user-friendlyness into the whole app. If the message service is offline, the page will render anyway and show a warning. Should also log to Sentry or other notification system. Might also run some daemon that can check service heartbeats, and sends a notification when a service doesn't reply.