Join user info with message
In the frontend, the user info is added to the database object, which can populate the user fields on the blog feed. On the frontpage, this is done in the backend. Would be nice to have one template snippet instead of jinja2 and Ractive-mustache version, but fast frontend rendering is important as well, and there are some differences. I guess any project gets messy after a while, and while refactoring is important, some messiness is tolerable.

Also put a coming-soon message instead of the map.

Also added a profile pic placeholder until I implement that functionality.

The blog was changed to always have media show in 16by8 fashion, with grey bars if more vertical resolution. Looks nice and shows more messages at once.