More robust cleanup
Had some error in attelegrambot where onDisconnect was called and the event loop was stopped, but cleanup wasn't called (maybe _session was no more) so the telegrambot thread just kept the process alive forever, throwing errors that the asyncio loop was closed. Maybe this will fix that particular situation and just exit so it can be restarted.