Livetrack gps point proper typing
Before sending off the dict with a gps_point, it needs to be properly typed in order to pass the json schema validation (and to be successfully inserted into the db of course).

Also added a *source* column to *gps_point* table that indicates where the point comes from, so we can label or filter.