Telegrambot bugfix and users.db refactor
Refactored users.db to work with a pool that is instantiated on class creation. Also had to fix the insert user method, that needs access to a transaction, which only exists on a connection. I basically kept the old way for that method, which is to pass in a connection when calling the method, and if that is not defined then getting a connection from the pool.