Fix dict comparison error
The priority queue uses tuples with a priority number as first item. This number can be equal to another, which then prompts the comparison of the second item, which is a dict and can't be compared. Fixed by introducing a unique count as second item. Can also be fixed by using some custom class, but this was simpler and faster.