Guess location of message
Using the location info of a user, the location of a message is guesstimated. It uses some simple rules, see `backend/location/main.py`. It tries to find the nearest points within an hour from the message timestamp, then linearly interpolates to the nearest points if there are both points before and after or chooses the closest point if there are only points before or after. If no points found within an hour, no message coordinates are set.