Adventure location tracks
Incoming location gps points are now emitted as updates on the applicable user and adventure channels.

It is also possible to retrieve all adventure tracks at once.

Fixed a small issue where the user_id was not passed along with the track in location.db, this is needed for the frontend.

Also refactored to use *at.public.* prefix for all public channels, both callable methods and subscribable topics.