Add sponsor logos to adventure
These logos are defined in a separate table with a foreign key to the adventure, their images being stored in the `/home/atuser/assets/logos/` folder.  If there are logos present, they are shown at the bottom in an automatic slider, on desktop and mobile. At some point the user might upload these logos in their adventure administration page (which also does not exist yet, doing everything manually in the database).