Tabs on mobile to switch between map/blog
The tab menu is only shown on mobile, and is a Ractive instance. On switching between map/blog, it adds the mobile-hide class to the other and removes it from the one that should be active. When switching to map, the map is resized, as for some reason the canvas is not properly resized to the parent element automatically (maybe too early or something).