Spotbot retrieves spot feeds
Feeds are linked to user_id, and spotbot will check them within the constraints of the SPOT API. It saves all messages internally to be able to identify new ones, and to be able to refer back to them. It sends the location of a track point or message to the location service.